Extract the ErrorBoundary to an independent file.
